Ensure `iterdir()` and `_scandir()` work when using multiple boards.
Save a reference to the `board` instance when creating a new `MPRemoteDirEntry`
instance. This allows the `stat()` method to work correctly with multiple
boards.

Initialise the `board` attribute of new `MPRemotePath` instances created by
`_scandir()` and `iterdir()`. This is a workaround because
`Path._make_child_relpath()` does not invoke the `MPRemotePath.__new__()`
method, which would normally set the `board` attribute.