Add github workflows for code checks, build, publish and release.
The workflows are based on the reusable workflows from the
glenn20/python-ci repository.

The pytest tests are ommitted as they require hardware-in-the-loop.
Run the CI tests on your local computer with a serial-attached ESP32
device using tox, eg: `tox -- --port=u0 -xv`, or
`uv run -p 3.12 pytest --port c3 -xv`.