tests: Add a unix micropython port for testing.
The unix port is used to simulate a serial port connection to the
micropython board. The micropython unix port is initialised with a
RAM filesystem as root fs to emulate a serial attached micropython
board.

Requires a micropython with support for the raw repl.

Requires the socat package to be installed on the system.

Enable the `socat` package for the CI workflow on github.

conftest.py: Refactor the test fixtures.

Enable all tests in CI workflow.